// console/src/api/generated/models/photo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Metadata } from './metadata';
// May contain unused imports in some cases
// @ts-ignore
import type { PhotoExif } from './photo-exif';
// May contain unused imports in some cases
// @ts-ignore
import type { PhotoSpec } from './photo-spec';

/**
 * 
 * @export
 * @interface Photo
 */
export interface Photo {
    /**
     * 
     * @type {string}
     * @memberof Photo
     */
    'apiVersion': string;
    /**
     * 
     * @type {PhotoExif}
     * @memberof Photo
     */
    'exif'?: PhotoExif;
    /**
     * 
     * @type {string}
     * @memberof Photo
     */
    'kind': string;
    /**
     * 
     * @type {Metadata}
     * @memberof Photo
     */
    'metadata': Metadata;
    /**
     * 
     * @type {PhotoSpec}
     * @memberof Photo
     */
    'spec': PhotoSpec;
}